Add Random Verse feature and update dependencies
