+++ extensions/bible-verse-search/package.json
@@ -27,12 +27,12 @@
   },
   "devDependencies": {
     "@raycast/eslint-config": "^1.0.11",
-    "@types/node": "22.10.2",
-    "@types/react": "19.0.2",
+    "@types/node": "^20.8.10",
+    "@types/react": "^18.3.3",
     "@typescript-eslint/eslint-plugin": "^8.19.0",
-    "eslint": "^9.17.0",
-    "prettier": "^3.4.2",
-    "typescript": "^5.7.2"
+    "eslint": "^8.57.0",
+    "react": "^18.2.0",
+    "typescript": "^5.4.5"
   },
   "scripts": {
     "build": "ray build",
